### 2022-11-14 #4893 @dankusanovic
Dear all,

I am available to review some papers. I would like to know the process on
how to carry this out,

Thanks for the consideration,

Danilo

On Mon, Nov 14, 2022 at 1:41 AM Gianluca Iori ***@***.***>
wrote:

> @danasolav yes sure.
> Maybe this people from the JOSS list?
> @jacobmerson
> @iammix
> @PolymerGuy
> @CWillberg
> @dankusanovic
> @abhishektha
> @melund
>
>
>
> On Mon, 14 Nov 2022 at 08:40, Dana Solav ***@***.***> wrote:
>
> > @gianthk <https://github.com/gianthk> if you have any suggestions for
> > potential reviewers then please mention them here in this thread (without
> > tagging them with an @). In addition, this list of people
> > <https://bit.ly/joss-reviewers> have already agreed to review for JOSS
> > and may be suitable for this submission (please start at the bottom of
> the
> > list).
> >
> > —
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/4893#issuecomment-1313167589
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/ADX6FOYQGX5UUCDKROOZLQ3WIHNGJANCNFSM6AAAAAARRVX7ME
> >
> > .
> > You are receiving this because you were mentioned.Message ID:
> > ***@***.***>
> >
>
