### 2022-11-21 #4893 @danasolav
@dankusanovic you can find the review guideline and checklist [here] (https://joss.readthedocs.io/en/latest/index.html). Once this review opens, you'll be directed to the review page, where you'll generate your checklist and work through it.
