### 2022-11-21 #4952 @editorialbot
[REVIEW]: Ciclope: micro Computed Tomography to Finite Elements
**Submitting author:** @gianthk (<a href="http://orcid.org/0000-0001-8611-3281">gianluca iori</a>)
**Repository:** https://github.com/gianthk/ciclope
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.4
