### 2022-12-14 gianthk/ciclope#9 @jacobmerson
missing tiff files from examples
Many of the examples reference tiff image stacks that are not available in the test_data folder.
