### 2022-12-14 #4952 @jacobmerson
@gianthk A few issues as I work through the review
1. Are there tests that can be run either automatically or manually (with documented steps?). I do not see this
2. Since your examples include data from humans and animals please review the policies here: https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies and add the appropriate statements to your paper and documentation.
3. You are missing data needed to reproduce any of your examples. See https://github.com/gianthk/ciclope/issues/9
