### 2022-12-14 #4952 @gianthk
>     1. Are there tests that can be run either automatically or manually (with documented steps?). I do not see this
@jacobmerson the examples cover all steps of software functionality. we will upload the test dataset to reproduce the examples. on top of this, are automated tests mandatory?
