### 2022-12-14 #4952 @jacobmerson
> @jacobmerson thank you for the revisions! This is my first JOSS paper. how does it work? can I start working on these already, commit changes and solve the issues you opened?

Yes! Please address any issues that we bring up through the review process as we go. One nice thing about JOSS is that things can be iterative.

> @jacobmerson the examples cover all steps of software functionality. we will upload the test dataset to reproduce the examples. on top of this, are automated tests mandatory?

Please take a look at the review criteria [here](https://joss.readthedocs.io/en/latest/review_criteria.html#tests). You do provide a lot of examples. However, based on my interpretation of the review criteria these examples alone are not enough. You need to have some mechanism for me to check the correctness of the outputs either manual or automated (preferred).  You don't necessarily need unit testing, but you do need integration tests.

In other words, given some set of input files and parameters, do I get the correct input files for Abaqus/Calculix out? As much of your public API as possible should have some form of testing.
