### 2022-12-16 #4952 @dankusanovic
@gianthk
I agree with @jacobmerson . In my opinion, the paper and examples read and execute well. The pip installation is quite nice. However, there is no test validation (automatic) if the elements are properly implemented. I suggest including a few that shows that Hexa, Tetra, Quad, and so on work well. On the same line, it will be nice to compare to some other software to see performance and validation of results. Abaqus and Ansys could help if a license is available, otherwise FreeFEM, FEniCs, Deal.II are opensource possibilities
