### 2023-01-14 a871368
removed unnecessary dependencies; updated dos and examples
