+++ docs/examples.rst
@@ -6,6 +6,10 @@ This section contains examples of **ciclope** pipelines.
 
 .. Examples can be run in your browser using the Binder link at the top of the page.
 
+The dataset used in the following examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: [FP6-IST 026932](https://cordis.europa.eu/project/id/026932).
+Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
+For info on the dataset see [here](http://projects.kmi.open.ac.uk/lhdl/).
+
 You can run all
 `Jupyter notebook <https://jupyter-notebook.readthedocs.io/en/stable/>`_
 examples on your local machine.
@@ -19,4 +23,4 @@ You can also execute the pipelines from the command line. The header of each exa
 
    ipynb/ciclope_ex01_voxeluFE_CalculiX
    ipynb/ciclope_ex02_tetrauFE_CalculiX
-   ipynb/ciclope_ex03_tetraFE_tooth_CalculiX
+..   ipynb/ciclope_ex03_tetraFE_tooth_CalculiX
