+++ src/ciclope/__init__.py
@@ -1,5 +1,5 @@
 """Computed Tomography to Finite Elements"""
-__version__ = "1.2.2"
+__version__ = "1.2.3"
 
 from .core import voxelFE, tetraFE
 # from .utils import preprocess, postprocess
