+++ docs/install.md
@@ -14,7 +14,8 @@ conda install -c conda-forge dxchange
 ## Testing
 Run tests to verify your installation with the command:
 ```commandline
-python -m unittest -v ciclope_tests.run_tests
+cd test
+python -m unittest -v test_ciclope.run_tests
 ```
 ---
 ## How to contribute
