+++ README.md
@@ -10,11 +10,11 @@ Computed Tomography to Finite Elements.
 
 ---
 ## Installation
-For mesh generation, `ciclope` requires [pygalmesh](https://github.com/meshpro/pygalmesh), a Python frontend to [CGAL](https://www.cgal.org/).
+For mesh generation, `ciclope` requires [pygalmesh-0.10.6](https://github.com/meshpro/pygalmesh), a Python frontend to [CGAL](https://www.cgal.org/).
 Follow the [installation procedure](https://github.com/meshpro/pygalmesh#installation) for [CGAL](https://www.cgal.org/) and [Eigen](http://eigen.tuxfamily.org/index.php?title=Main_Page).
 After that install pygalmesh with:
 ```commandline
-conda install -c conda-forge pygalmesh
+conda install -c conda-forge pygalmesh==0.10.6
 ```
 After installing pygalmesh, you can install `ciclope` using pip. The flag `[all]` will install optional dependencies needed to run full pipelines and examples.
 ```commandline
