### 2023-01-15 7257d10
added ethics statement LHDL project
