+++ JOSS_2022/paper.md
@@ -84,7 +84,10 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 [comment]: <> (Dxchange @decarlo_2014)
 
 # Examples
-`Ciclope` contains a library of `Jupyter` notebooks of example applications in the field of computational biomechanics (\autoref{fig:examples}). The main use case is a pipeline for the generation of microFE models from microCT scans of trabecular bone (\autoref{fig:examples}A). The simulation (linear-elastic) of a mechanical compression test is used to back-calculate the apparent elastic modulus of trabecular bone. This procedure is demonstrated using hexahedra (voxel, \autoref{fig:examples}B), and tetrahedra (\autoref{fig:examples}C) finite elements. Two approaches for the local mapping of material inhomogeneities are illustrated using voxel and tetrahedra FE. Each example can be run within `Jupyter` or executed from the commandline with the `ciclope.py` script. `Ciclope` can be applied to microCT scans other than trabecular bone, such as whole teeth, or metal foams.
+`Ciclope` contains a library of `Jupyter` notebooks of example applications in the field of computational biomechanics (\autoref{fig:examples}). The main use case is a pipeline for the generation of microFE models from microCT scans of trabecular bone (\autoref{fig:examples}A). The dataset used in the examples is part of the public collection of the Living Human Digital Library ([LHDL](http://projects.kmi.open.ac.uk/lhdl/)) Project, a project financed by the European Commission (project number: [FP6-IST 026932](https://cordis.europa.eu/project/id/026932).
+Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
+
+A linear-elastic simulation of a mechanical compression test is used to back-calculate the apparent elastic modulus of trabecular bone. This procedure is demonstrated using hexahedra (voxel, \autoref{fig:examples}B), and tetrahedra (\autoref{fig:examples}C) finite elements. Two approaches for the local mapping of material inhomogeneities are illustrated using voxel and tetrahedra FE. Each example can be run within `Jupyter` or executed from the commandline with the `ciclope.py` script. `Ciclope` can be applied to microCT scans other than trabecular bone, such as whole teeth, or metal foams.
 
 ![MicroFE models of trabecular bone generated from 3D microCT images with `ciclope`. (A) Input microCT volume data. (B) Hexahedra, and (C) tetrahedra finite element models generated with the `voxelFE.py` and `tetraFE.py` modules, respectively.\label{fig:examples}](examples.png){width=100%}
 
