+++ docs/examples.rst
@@ -6,9 +6,9 @@ This section contains examples of **ciclope** pipelines.
 
 .. Examples can be run in your browser using the Binder link at the top of the page.
 
-The dataset used in the following examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: [FP6-IST 026932](https://cordis.europa.eu/project/id/026932).
+The dataset used in the following examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: `FP6-IST 026932 <https://cordis.europa.eu/project/id/026932>`).
 Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
-For info on the dataset see [here](http://projects.kmi.open.ac.uk/lhdl/).
+For info on the dataset see `here <http://projects.kmi.open.ac.uk/lhdl/>`.
 
 You can run all
 `Jupyter notebook <https://jupyter-notebook.readthedocs.io/en/stable/>`_
