+++ README.md
@@ -8,7 +8,7 @@ Computed Tomography to Finite Elements.
 
 **ciclope** processes micro Computed Tomography (microCT) data to generate Finite Element (FE) models. <br />
 
----
+
 ## Installation
 For mesh generation, `ciclope` requires [pygalmesh-0.10.6](https://github.com/meshpro/pygalmesh), a Python frontend to [CGAL](https://www.cgal.org/).
 Follow the [installation procedure](https://github.com/meshpro/pygalmesh#installation) for [CGAL](https://www.cgal.org/) and [Eigen](http://eigen.tuxfamily.org/index.php?title=Main_Page).
@@ -34,7 +34,7 @@ python -m unittest -v test_ciclope.run_tests
 ### How to contribute
 If you want to contribute to this project, please install `ciclope` following the [development guide](development.md).
 
----
+
 ## Usage
 **ciclope** pipelines can be run from the command line as a script. Scroll down and take a look at the [Examples](###Examples) folder for this type of use.
 To view the command line script help run:
@@ -126,11 +126,11 @@ ciclope.utils.postprocess.paraview_plot("test_data/tooth/results/Tooth_3_scaled_
 |:-------------------------:|:-------------------------:|:-------------------------:|
 |![](test_data/tooth/results/Tooth_3_scaled_2_S_Mises_XY.png) | ![](test_data/tooth/results/Tooth_3_scaled_2_S_Mises_XZ.png) | ![](test_data/tooth/results/Tooth_3_scaled_2_S_Mises_YZ.png) |
 
----
-## ciclope pipeline 
-The following table shows a general pipeline for FE model generation from CT data that can be executed with ciclope:
 
-| # | Step | Description | **ciclope** flag |
+## ciclope pipeline 
+The following table shows a general pipeline for FE model generation from CT data that can be executed from the command line with the `ciclope` script:
+
+| # | Step | Description | command line script flag |
 |:-:|:-|:-|:-|
 | 1. | **Load CT data** | | |
 | 2. | **Pre-processing** | Gaussian smooth | `--smooth` |
@@ -146,14 +146,13 @@ The following table shows a general pipeline for FE model generation from CT dat
 | | | Voxel FE | `--voxelfe` |
 | | | Tetrahedra FE | `--tetrafe` |
 
----
+
 ## Notes on ciclope
 * Tetrahedra meshes are generated with [pygalmesh](https://github.com/nschloe/pygalmesh) (a Python frontend to [CGAL](https://www.cgal.org/))
 * High-resolution surface meshes for visualization are generated with the [PyMCubes](https://github.com/pmneila/PyMCubes) module.
 * All mesh exports are performed with the [meshio](https://github.com/nschloe/meshio) module.
 * **ciclope** handles the definition of material properties and FE analysis parameters (e.g. boundary conditions, simulation steps..) through separate template files. The folders [material_properties](/material_properties) and [input_templates](/input_templates) contain a library of template files that can be used to generate FE simulations.
   * Additional libraries of [CalculiX](https://github.com/calculix) examples and template files can be found [here](https://github.com/calculix/examples) and [here](https://github.com/calculix/mkraska)
-___
 
 ## Examples
[…]
