### 2023-01-15 aa19a67
run full NB example 01
