+++ src/ciclope/__about__.py
@@ -1,0 +1,13 @@
+try:
+    # Python 3.8+
+    from importlib import metadata
+except ImportError:
+    try:
+        import importlib_metadata as metadata
+    except ImportError:
+        __version__ = "unknown"
+
+try:
+    __version__ = metadata.version("ciclope")
+except Exception:
+    __version__ = "unknown"
