### 2023-01-18 18ddd57
updated example NB narrative and dirs
