### 2023-01-20 aa63237
updated example NBs and description
