+++ docs/examples.rst
@@ -6,9 +6,9 @@ This section contains examples of **ciclope** pipelines.
 
 .. Examples can be run in your browser using the Binder link at the top of the page.
 
-The dataset used in the following examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: `FP6-IST 026932 <https://cordis.europa.eu/project/id/026932>`).
+The trabecular bone micrCT dataset used in the examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: `FP6-IST 026932 <https://cordis.europa.eu/project/id/026932>`_).
 Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
-For info on the dataset see `here <http://projects.kmi.open.ac.uk/lhdl/>`.
+For info on the dataset see `here <http://projects.kmi.open.ac.uk/lhdl/>`_.
 
 You can run all
 `Jupyter notebook <https://jupyter-notebook.readthedocs.io/en/stable/>`_
@@ -23,5 +23,5 @@ You can also execute the pipelines from the command line. The header of each exa
 
    ipynb/ciclope_ex01_voxeluFE_CalculiX
    ipynb/ciclope_ex02_tetrauFE_CalculiX
-..   ipynb/ciclope_ex03_tetraFE_tooth_CalculiX
-..   ipynb/ciclope_ex04_tetraFE_steelfoam_nonlinear_CalculiX
+   ipynb/ciclope_ex03_tetraFE_tooth_CalculiX
+   ipynb/ciclope_ex04_tetraFE_steelfoam_nonlinear_CalculiX
