+++ src/ciclope/__init__.py
@@ -1,5 +1,5 @@
 """Computed Tomography to Finite Elements"""
-__version__ = "1.3.0"
+__version__ = "1.3.1"
 
 from .core import voxelFE, tetraFE
 # from .utils import preprocess, postprocess
