+++ src/ciclope/utils/postprocess.py
@@ -4,9 +4,6 @@
 Ciclope postprocessing module
 """
 
-import sys
-sys.path.append('~/Applications/ParaView-5.9.0-RC1-MPI-Linux-Python3.8-64bit/lib/python3.8/site-packages')
-
 try:
     from paraview.simple import *
 except ImportError:
