### 2023-01-20 #4952 @gianthk
@jacobmerson @dankusanovic here we go thank you very much for reviewing all ciclope functionalities! It took us some time but it was a chance to find and debug more issues with installation.

- [X] we fixed the non-optional list of dependencies in `project.toml`
- [X] we discovered issues with the versions of `pygalmesh` and `meshio`. now `pip install ciclope` forces the installation of the correct versions: `pygalmesh==0.10.6` and `meshio==5.0.0`
- [X] @jacobmerson we agree that CGAL must be installed as a requirement but pycgal is not! we don't know why you had install it. The sequence is CGAL and Eigen, then pygalmesh (and ciclope). We added these instructions with links to the [readthedocs](https://ciclope.readthedocs.io/en/latest/install.html) and [github readme](https://github.com/gianthk/ciclope#installation) pages.

### unit testing
- [X] we uploaded a [small dataset](https://github.com/gianthk/ciclope/tree/master/test/trab) (10x10x10 image) for unit tests
- [X] unit tests implemented for the key methods of `ciclope.core`:
    1. voxelFE.vol2ugrid
    2. voxelFE.mesh2voxelfe
    3. tetraFE.shell_mesh (both `pymcubes` and `pygalmesh` methods
    4. tetraFE.cgal_mesh
    5. tetraFE.mesh2tetrafe

- [X] How to perform unit tests described in the [readme](https://github.com/gianthk/ciclope#testing) as well as [readthedocs](https://ciclope.readthedocs.io/en/latest/install.html#testing)

- For both `voxelFE` and `tetraFE`, the outputs of mesh generation are asserted versus element types and non-empty mesh fields. Similarly, FE input files generated on the synthetic 10x10x10 dataset are checked to assert existence of FE keywords (nodes, elements, analysis definition..). In this way we verify that the written FE files are correct.
- @dankusanovic ciclope is not and does not provide an FE solver. We think that the testing and comparison of different FE solvers, or of open-source vs proprietary solvers is out of the scope of our work. We provide unit tests for the functionality of ciclope (generation of FE input files)

### test_data for examples and tests
- [X] we uploaded the LHDL trabecular bone dataset to [test_data](https://github.com/gianthk/ciclope/tree/master/test_data/LHDL/3155_D_4_bc/cropped)
- [X] we uploaded a [small dataset](https://github.com/gianthk/ciclope/tree/master/test/trab) (10x10x10 image) for unit tests
- [X] we added the ethics statement of the LHDL Project to the example [notebooks](https://github.com/gianthk/ciclope/blob/master/examples/ipynb/ciclope_ex01_voxeluFE_CalculiX.ipynb), [readthedocs](https://ciclope.readthedocs.io/en/latest/examples.html), and paper

**the new release v1.3.1 should incorporate all the items discussed in the review.**
