+++ JOSS_2022/paper.bib
@@ -143,7 +143,7 @@
 	copyright = {Copyright © 2017 John Wiley \& Sons, Ltd.},
 	issn = {2040-7947},
 	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cnm.2941},
-	doi = {https://doi.org/10.1002/cnm.2941},
+	doi = {10.1002/cnm.2941},
 	abstract = {In this study, the sensitivity of the apparent response of trabecular bone to different constitutive models at the tissue level was investigated using finite element (FE) modeling based on micro-computed tomography (micro-CT). Trabecular bone specimens from porcine femurs were loaded under a uniaxial compression experimentally and computationally. The apparent behaviors computed using von Mises, Drucker-Prager, and Cast Iron plasticity models were compared. Secondly, the effect of bone volume fraction was studied by changing the bone volume fraction of a trabecular bone sample while keeping the same basic architecture. Also, constitutive models' parameters of the tissue were calibrated for porcine bone, and the effects of different parameters on resulting apparent response were investigated through a parametric study. The calibrated effective tissue elastic modulus of porcine trabecular bone was 10±1.2 GPa, which is in the lower range of modulus values reported in the literature for human and bovine trabecular bones (4–23.8 GPa). It was also observed that, unlike elastic modulus, yield properties of tissue could not be uniquely calibrated by fitting an apparent response from simulations to experiments under a uniaxial compression. Our results demonstrated that using these 3 tissue constitutive models had only a slight effect on the apparent response. As expected, there was a significant change in the apparent response with varying bone volume fraction. Also, both apparent modulus and maximum stress had a linear relation with bone volume fraction.},
 	language = {en},
 	number = {4},
