+++ docs/examples.rst
@@ -6,9 +6,13 @@ This section contains examples of **ciclope** pipelines.
 
 .. Examples can be run in your browser using the Binder link at the top of the page.
 
-The trabecular bone micrCT dataset used in the examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number: `FP6-IST 026932 <https://cordis.europa.eu/project/id/026932>`_).
+The trabecular bone micrCT dataset used in the examples is part of the public collection of the Living Human Digital Library (LHDL) Project, a project financed by the European Commission (project number:
+`FP6-IST 026932 <https://cordis.europa.eu/project/id/026932>`_
+).
 Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
-For info on the dataset see `here <http://projects.kmi.open.ac.uk/lhdl/>`_.
+For info on the dataset see
+`here <http://projects.kmi.open.ac.uk/lhdl/>`_
+.
 
 You can run all
 `Jupyter notebook <https://jupyter-notebook.readthedocs.io/en/stable/>`_
