### 2023-02-05 gianthk/ciclope#14 @jacobmerson
statement for animal/human samples for JOSS paper
As I mentioned in the JOSS review thread, you need to add statements about using data from animals/humans:
https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies
