### 2023-02-05 #4952 @gianthk
>     * [ ]  Update the paper with the appropriate statements for human/animal testing? [statement for animal/human samples for JOSS paper gianthk/ciclope#14](https://github.com/gianthk/ciclope/issues/14)
Yes we did. The latest paper version is in [this Action's run](https://github.com/gianthk/ciclope/actions/runs/3981147994).

> The microCT bone dataset used in the examples is part of the public collection of the Living Human Digital Library (LHDL) project, a project financed by the European Commission (project number: FP6-IST 026932). Human tissues in the LHDL project were collected according to the body donation program of Universitè Libre de Bruxelles (ULB), partner of the project.
