### 2023-02-05 #4952 @jacobmerson
@danasolav can you help me understand if the statements made by the authors about human/animal samples are adequate for JOSS?

They provide a reference to the public access database that they retrieved the imaging samples from (and did not collect the samples themselves). However, the JOSS guidelines require
> The manuscript must include complete reporting of the data, including the ways that the study was approved and relevant details of the sample. Authors are required to report either the [ARRIVE](https://arriveguidelines.org/) or [PREPARE](https://doi.org/10.1177/0023677217724823) guidelines for animal research in the submission.

However they do not mention the `ARRIVE` or `PREPARE` guidelines. From a pragmatic standpoint it seems that referencing the public database is adequate since they do not collect the samples themselves.
