### 2023-02-05 gianthk/ciclope#14 @jacobmerson
The references to the dataset in the updated paper seem fine to me from a pragmatic standpoint (since you don't collect any samples yourselves and are using a public database). However, it still doesn't meet my reading of the strict requirements in the JOSS webpage.

> The manuscript must include complete reporting of the data, including the ways that the study was approved and relevant details of the sample. Authors are required to report either the [ARRIVE](https://arriveguidelines.org/) or [PREPARE](https://doi.org/10.1177/0023677217724823) guidelines for animal research in the submission.

I have asked the editor for more clarification on the issue.
