### 2023-02-05 gianthk/ciclope#14 @jacobmerson
@gianthk I re-read the guidelines and you are all clear on this. It is now clear to me that you have not included "original data on animal research".
