### 2023-02-05 #4952 @jacobmerson
@gianthk @danasolav @dankusanovic Just went through the JOSS requirements again and it does specify "original" research needs to include those guidelines however, with their paper edits it is clear that this is not original animal/human research.

They should be all set.
