+++ JOSS_2022/paper.md
@@ -76,10 +76,10 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 `Ciclope` relies on several other tools for 3D image and FE processing:
 
 - Voxel and tetrahedra mesh exports are performed with `meshio` [@schlomer_meshio_2022].
-- Tetrahedra meshes are generated with `pygalmesh`, a Python frontend to [`CGAL`](https://www.cgal.org/) [@schlomer_pygalmesh_2022].
-- High-resolution surface meshes for visualization are generated with `PyMCubes` [@neila_pymcubes_2022].
+- Tetrahedra meshes are generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
+- High-resolution surface meshes are generated with `PyMCubes` [@neila_pymcubes_2022].
 - The FE input files (`.INP`) generated by `ciclope` can be solved using the free software `CalculiX` [@dhont_calculix_2022] or [`Abaqus`](https://www.3ds.com/products-services/simulia/products/abaqus/).
-- For data and FE results visualization, `ciclope` relies on `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022].
+- 3D images and FE results are visualized with `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022].
 
 [comment]: <> (Dxchange @decarlo_2014)
 
