+++ JOSS_2022/paper.bib
@@ -464,7 +464,7 @@ Publisher: International Union of Crystallography},
 }
 
 @misc{noauthor_calculix_2022,
-	title = {{CalculiX} to {Paraview} converter (frd to vtk/vtu)},
+	title = {{CalculiX} to {Paraview} converter},
 	copyright = {GPL-3.0},
 	url = {https://github.com/calculix/ccx2paraview},
 	abstract = {CalculiX to Paraview converter (frd to vtk/vtu). Makes possible to view and postprocess CalculiX analysis results in Paraview. Generates Mises and Principal components for stress and strain tensors.},
