+++ pyproject.toml
@@ -37,7 +37,7 @@ dynamic = ["version"]
 dependencies = [
     "numpy",
     "meshio==5.0.0",
-    "pygalmesh==0.10.6",
+    "pygalmesh>=0.10.6",
     "PyMCubes",
     "scipy",
     "scikit-image",
