### 2023-02-05 #4952 @jacobmerson
@gianthk A few comments on the paper. I don't consider these to be blockers, but they might improve the quality of your paper:

Line 33-34: you may want to indicate how standard microCT datasets can be converted to numpy arrays. This discussion may be too much for the paper, but you should have this in the documentation. Or, at least a description of the structure of the expected inputs. A `t x n x n` array where `t` the number of frames through the thickness and `n` is the number of pixels across each edge.
Line 45: you should say what a `.inp` file is. i.e., say that it is the input file format for Abaqus/Calculix
Line 50: why is material mapping bolded?
Line 56-65: You mention that there are components you rely on for visualization, but have not included any description of what visualization capabilities exist within Ciclope
