### 2023-02-05 #4952 @jacobmerson
@gianthk @danasolav @dankusanovic My review is complete. I would like to recommend this work for publication. This paper provides open source tools to aid in the construction of finite element models from microCT image stack data including basic tools for segmentation, as well as the construction of input files for Abaqus and Calculix. These types of pipelines are challenging to construct, but lay the foundation for important advancements in the use of patient specific data in research and clinical settings.

I have suggested some minimal clarification edits to the author, however I don't consider these to be required.
