+++ JOSS_2022/paper.md
@@ -66,7 +66,7 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 **Meshing**: `ciclope` allows to create two types of FE meshes. Image voxels can be directly converted to 8-node, hexahedral brick elements with the `voxelFE.py` module. Alternatively, meshes of 4-node tetrahedra can be generated calling `CGAL` [@the_cgal_project_cgal_2022] through the `tetraFE.py` module. 
 
 [comment]: <> (Finally, the 'beamFE.py' module allows to generate a mesh of 2-node beam elements, where each beam represents a single trabecula, and has a local trabecular thickness associated to it.)
-**FE model generation**: the mesh is converted to an `.INP` input file for the FE solver. Within model generation, it is possible to assign material properties, define the boundary conditions, analysis type and steps, and request specific simulation outputs using a separate analysis template `.TMP` file. Libraries of `material_properties` and `input_templates` are provided for this process. Additional `CalculiX` user examples and templates are available online [@noauthor_calculix_2022-1]. For voxel-FE model generation, different **material mapping** strategies can be used: uniform tissue material properties (elastic modulus and poisson ratio) can be applied to all bone voxels. Alternatively, the local image intensity (voxel grey values) can be converted to heterogeneous material properties using a mapping law defined by the user.
+**FE model generation**: the mesh is converted to an `.INP` input file for the FE solver. Within model generation, it is possible to assign material properties, define the boundary conditions, analysis type and steps, and request specific simulation outputs using a separate analysis template `.TMP` file. Libraries of `material_properties` and `input_templates` are provided for this process. Additional `CalculiX` user examples and templates are available online [@noauthor_calculix_2022-1]. For voxel-FE model generation, different material mapping strategies can be used: uniform tissue material properties (elastic modulus and poisson ratio) can be applied to all bone voxels. Alternatively, the local image intensity (voxel grey values) can be converted to heterogeneous material properties using a mapping law defined by the user.
 
 [comment]: <> (bone mineral density BMD through a calibration rule obtained scanning a hydroxyapatite phantom. After this, an empirical law is used to convert local BMD to tissue elastic moduli Bourne_2004; garcia_2008.)
 [comment]: <> (The pipeline is composed of the following steps:)
