### 2023-02-07 #4952 @gianthk
> Line 33-34: you may want to indicate how standard microCT datasets can be converted to numpy arrays
- [ ]

> Line 45: you should say what a `.inp` file is. i.e., say that it is the input file format for Abaqus/Calculix
- [X] done

> Line 50: why is material mapping bolded?
- [X] bold removed

> Line 56-65: You mention that there are components you rely on for visualization, but have not included any description of what visualization capabilities exist within Ciclope
- [ ]
