+++ JOSS_2022/paper.md
@@ -66,7 +66,7 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 **Meshing**: `ciclope` allows to create two types of FE meshes. Image voxels can be directly converted to 8-node, hexahedral brick elements with the `voxelFE.py` module. Alternatively, meshes of 4-node tetrahedra can be generated calling `CGAL` [@the_cgal_project_cgal_2022] through the `tetraFE.py` module. 
 
 [comment]: <> (Finally, the 'beamFE.py' module allows to generate a mesh of 2-node beam elements, where each beam represents a single trabecula, and has a local trabecular thickness associated to it.)
-**FE model generation**: the mesh is converted to an `.INP` input file for the FE solver. Within model generation, it is possible to assign material properties, define the boundary conditions, analysis type and steps, and request specific simulation outputs using a separate analysis template `.TMP` file. Libraries of `material_properties` and `input_templates` are provided for this process. Additional `CalculiX` user examples and templates are available online [@noauthor_calculix_2022-1]. For voxel-FE model generation, different **material mapping** strategies can be used: uniform tissue material properties (elastic modulus and poisson ratio) can be applied to all bone voxels. Alternatively, the local image intensity (voxel grey values) can be converted to heterogeneous material properties using a mapping law defined by the user.
+**FE model generation**: the mesh is converted to an `.INP` input file for Abaqus/CalculiX FE solvers. Within model generation, it is possible to assign material properties, define the boundary conditions, analysis type and steps, and request specific simulation outputs using a separate analysis template `.TMP` file. Libraries of `material_properties` and `input_templates` are provided for this process. Additional `CalculiX` user examples and templates are available online [@noauthor_calculix_2022-1]. For voxel-FE model generation, different **material mapping** strategies can be used: uniform tissue material properties (elastic modulus and poisson ratio) can be applied to all bone voxels. Alternatively, the local image intensity (voxel grey values) can be converted to heterogeneous material properties using a mapping law defined by the user.
 
 [comment]: <> (bone mineral density BMD through a calibration rule obtained scanning a hydroxyapatite phantom. After this, an empirical law is used to convert local BMD to tissue elastic moduli Bourne_2004; garcia_2008.)
 [comment]: <> (The pipeline is composed of the following steps:)
@@ -75,11 +75,11 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 ### The ciclope ecosystem
 `Ciclope` relies on several other tools for 3D image and FE processing:
 
-- Voxel and tetrahedra mesh exports are performed with `meshio` [@schlomer_meshio_2022].
-- Tetrahedra meshes are generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
-- High-resolution surface meshes are generated with `PyMCubes` [@neila_pymcubes_2022].
-- The FE input files (`.INP`) generated by `ciclope` can be solved using the free software `CalculiX` [@dhont_calculix_2022] or [`Abaqus`](https://www.3ds.com/products-services/simulia/products/abaqus/).
-- 3D images and FE results are visualized with `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022].
+- Voxel and tetrahedra mesh exports performed with `meshio` [@schlomer_meshio_2022].
+- Tetrahedra meshes generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
[…]
