+++ docs/usage.md
@@ -1,5 +1,5 @@
 # Usage
-A general pipeline for FE model generation from microCT data is composed of the following steps:
+A general pipeline for FE model generation from microCT data is composed of the following steps
 ![FEM pipeline](./../examples/fem_pipeline.png)
 
 To use **ciclope** within python, import the package with
@@ -8,17 +8,20 @@ import ciclope
 ```
 ---
 ## Image pre-processing
-`ciclope.utils` contains functions that help you read and pre-process 3D datasets for FE model generation. You can skip this and use `ciclope` with 3D data as [`numpy.ndarray`](https://numpy.org/doc/stable/reference/generated/numpy.ndarray.html), or directly with a 3D [`meshio`](https://github.com/nschloe/meshio) mesh object.
+`ciclope.utils` contains functions that help you read and pre-process 3D datasets for FE model generation.
 
-Read 3D CT dataset stored as stack of TIFFs:
+Read 3D CT dataset stored as stack of TIFFs
 ```python
 from ciclope.utils.recon_utils import read_tiff_stack
 
-input_file = '/path_to_your_file/slice_0000.tiff'
+input_file = './test_data/LHDL/3155_D_4_bc/cropped/3155_D_4_bc_0000.tif'
 
 data_3D = read_tiff_stack(input_file)
 vs = np.ones(3) * 0.06  # voxelsize [mm]
 ```
+`read_tiff_stack` reads all TIFF files (slices) contained in the `input_file` folder. The volume is stored in a 3D [`numpy.ndarray`](https://numpy.org/doc/stable/reference/generated/numpy.ndarray.html) with size `[slices, rows, columns]`.
+
+---
 Segment the 3D image and remove unconnected voxels:
 ```python
 from skimage import morphology
@@ -29,39 +32,37 @@ BW = morphology.closing(BW, morphology.ball(2)) # optional step
 L = remove_unconnected(BW)
 ```
 ---
-## FE model generation
+## Mesh and FE model generation
+If you already have a mesh file, you can skip the mesh generation steps and use `ciclope` with 3D [`meshio`](https://github.com/nschloe/meshio) objects.
 ### voxel-FE
 ![](./../test_data/trabecular_bone/trab_sample_mini3_UD3.png)
-Generate **voxel-FE** model of linear elastic compression test:
+Generate unstructured grid mesh of hexahedra (voxels)
 ```python
 import ciclope
-
-input_template = "./input_templates/tmp_example01_comp_static_bone.inp"
-
-# generate unstructured grid mesh
 mesh = ciclope.core.voxelFE.vol2ugrid(data_3D, vs)
+```
 
-# generate CalculiX input file
+Generate CalculiX input file `.INP` for **voxel-FE** model of linear elastic compression test
+```python
+input_template = "./input_templates/tmp_example01_comp_static_bone.inp"
 ciclope.core.voxelFE.mesh2voxelfe(mesh, input_template, 'foo.inp', keywords=['NSET', 'ELSET'])
 ```
 
 ### tetrahedra-FE
 ![](./../test_data/steel_foam/B_matrix_tetraFE_mesh.png)
-Generate mesh of tetrahedra. `ciclope` uses [`pygalmesh`](https://github.com/nschloe/pygalmesh) for tetrahedra mesh generation:
+Generate mesh of tetrahedra. `ciclope` uses [`pygalmesh`](https://github.com/nschloe/pygalmesh) for tetrahedra mesh generation
 ```python
 mesh = ciclope.core.tetraFE.cgal_mesh(L, vs, 'tetra', max_facet_distance=0.2, max_cell_circumradius=0.1)
 ```
 
[…]

[…]
