+++ JOSS_2022/paper.md
@@ -75,11 +75,11 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 ### The ciclope ecosystem
 `Ciclope` relies on several other tools for 3D image and FE processing:
 
-- Voxel and tetrahedra mesh exports are performed with `meshio` [@schlomer_meshio_2022].
-- Tetrahedra meshes are generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
-- High-resolution surface meshes are generated with `PyMCubes` [@neila_pymcubes_2022].
-- The FE input files (`.INP`) generated by `ciclope` can be solved using the free software `CalculiX` [@dhont_calculix_2022] or [`Abaqus`](https://www.3ds.com/products-services/simulia/products/abaqus/).
-- 3D images and FE results are visualized with `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022].
+- Voxel and tetrahedra mesh exports performed with `meshio` [@schlomer_meshio_2022].
+- Tetrahedra meshes generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
+- High-resolution surface meshes generated with `PyMCubes` [@neila_pymcubes_2022].
+- FE input files (`.INP`) generated by `ciclope` can be solved using the free software `CalculiX` [@dhont_calculix_2022] or [`Abaqus`](https://www.3ds.com/products-services/simulia/products/abaqus/).
+- 3D images and FE results can be visualized with `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022] as illustrated in the [example Jupyter notebooks](#Examples).
 
 [comment]: <> (Dxchange @decarlo_2014)
 
