### 2023-02-08 d546cd3
updated install descriptions
updated requirements in install procedure
