### 2023-02-09 #4952 @danasolav
@dankusanovic please check your last unchecked box:
Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? **If the paper contains no such data, please check this item**.
