### 2023-02-14 gianthk/ciclope#16 @danasolav
Paper review
I have a few comments/suggestions about the current version of the paper:
- [ ] Line 20: is the package limited to isotropic linear elasticity? If not, please describe what other material models are available.
- [ ] Lines 32-54 describing the pipeline elements: the first two (image preprocessing and meshing) is in the same paragraph and the third is in a new paragraph. Please make it consistent. Also, please add a description for FE model solution and results postprocessing as well.
- [ ] Line 35: what is Otsu's method? I don't see it mentioned in the reference (Walt 2014).
- [ ] Line 36: should be "clusters of voxels"
- [ ] Please add the references related to the LHDL project as regular references in the list, and not only as links.
- [ ] Line 71: The word "Project" after (LHDL) can be deleted.
- [ ] Kine 73: should be "a partner of this project".
- [ ] Line 74: the hyphen in linear-elastic can be removed.
- [ ] Line 75: what about the Poisson ratio? can it also be calculated?
- [ ] What other examples are included besides the described main application?
- [ ] Please check all the cited software, packages, and repositories if they have citable publications other than the web address.
