+++ JOSS_2022/paper.bib
@@ -623,3 +623,19 @@ Publisher: International Union of Crystallography},
 	file = {Full Text PDF:/home/gianthk/Zotero/storage/HHLK6ZJK/Walt et al. - 2014 - scikit-image image processing in Python.pdf:application/pdf;Snapshot:/home/gianthk/Zotero/storage/S2SUMRD2/453.html:text/html},
 }
 
+@article{otsu_threshold_1979,
+	title = {A threshold selection method from gray-level histograms},
+	volume = {9},
+	number = {1},
+	journal = {IEEE Transactions on systems, man, and cybernetics},
+	author = {Otsu, Nobuyuki},
+	year = {1979},
+	pages = {62--66},
+}
+
+
+@misc{LHDL,
+	title = {The {Living} {Human} {Digital} {Library}},
+	url = {http://projects.kmi.open.ac.uk/lhdl/},
+	urldate = {2023-02-20},
+}
