+++ JOSS_2022/paper.md
@@ -62,12 +62,12 @@ We present the Python package `ciclope`: a fully open-source pipeline from micro
 ![Design of ciclope, and application to a pipeline for FE model generation from microCT data.\label{fig:design}](./../docs/ciclope_design.png){width=100%}
 
 A pipeline for the generation and solution of a FE model derived from 3D microCT data is shown in the central part of \autoref{fig:design}.
-**Image preprocessing**: a microCT dataset is loaded as a `NumPy` `ndarray` within Python and segmented to isolate bone voxels and background using the Otsu's method provided by `scikit-image` [@walt_scikit-image_2014]. A connectivity check is performed to remove isolated cluster of voxels, that the segmentation might have disconnected from the main structure. According to the user's needs, additional image processing can be applied for smoothing, cropping, resampling and rotating the dataset using the `SciPy` [@virtanen_scipy_2020] and `scikit-image` Python libraries. Furthermore, additional layers of material can be added at the top and bottom surfaces of the sample should the user need to replicate the effect of endcaps in the actual mechanical testing conditions.
-**Meshing**: `ciclope` allows to create two types of FE meshes. Image voxels can be directly converted to 8-node, hexahedral brick elements with the `voxelFE.py` module. Alternatively, meshes of 4-node tetrahedra can be generated calling `CGAL` [@the_cgal_project_cgal_2022] through the `tetraFE.py` module. 
+**Image preprocessing**: a microCT dataset is loaded as a `NumPy` `ndarray` within Python and segmented to isolate bone voxels and background using Otsu's method [@otsu_threshold_1979] as provided by `scikit-image` [@walt_scikit-image_2014]. A connectivity check is performed to remove isolated clusters of voxels, that the segmentation might have disconnected from the main structure. According to the user's needs, additional image processing can be applied for smoothing, cropping, resampling and rotating the dataset using the `SciPy` [@virtanen_scipy_2020] and `scikit-image` Python libraries. Furthermore, additional layers of material can be added at the top and bottom surfaces of the sample should the user need to replicate the effect of endcaps in the actual mechanical testing conditions.
+**Meshing**: `ciclope` allows to create two types of FE meshes. Image voxels can be directly converted to 8-node, hexahedral brick elements with the `voxelFE.py` module. Alternatively, meshes of 4-node tetrahedra can be generated calling `CGAL` [@the_cgal_project_cgal_2022] through the `tetraFE.py` module.
+**FE model generation**: the mesh is converted to an `.INP` input file for `Abaqus/CalculiX` FE solvers. Within model generation, it is possible to assign material properties, define the boundary conditions, analysis type and steps, and request specific simulation outputs using a separate analysis template `.TMP` file. Libraries of `material_properties` and `input_templates` are provided for this process. Additional `CalculiX` user examples and templates are available online [@noauthor_calculix_2022-1]. For voxel-FE model generation, different material mapping strategies can be used: uniform tissue material properties (elastic modulus and poisson ratio) can be applied to all bone voxels. Alternatively, the local image intensity (voxel grey values) can be converted to heterogeneous material properties using a mapping law defined by the user.
+**FE model solution and postprocessing**: FE models can be solved using the external software `Abaqus` or `CalculiX`. Simulation ouptut files are read to compute, among other, total reaction forces on the model boundaries, or orthogonal cross-section plots of the model's displacement of stress fields.
 
 [comment]: <> (Finally, the 'beamFE.py' module allows to generate a mesh of 2-node beam elements, where each beam represents a single trabecula, and has a local trabecular thickness associated to it.)
[…]
