### 2023-02-20 gianthk/ciclope#16 @gianthk
> * ck all the cited software, packages, and repositories if they have citable publications other than the web address.
