+++ JOSS_2022/paper.bib
@@ -411,9 +411,9 @@ Section: New Results},
 }
 
 @misc{schlomer_meshio_2022,
-	title = {meshio: {Tools} for mesh files},
+	title = {meshio v5.0.0: {Tools} for mesh files},
 	url = {https://github.com/nschloe/meshio},
-	author = {Schlömer, Nico},
+	author = {Schlömer, Nico, et al.},
 	year = {2022},
 	doi = {10.5281/zenodo.1173115},
 }
@@ -421,7 +421,7 @@ Section: New Results},
 @misc{schlomer_pygalmesh_2022,
 	title = {pygalmesh: {Python} interface for {CGAL}'s meshing tools},
 	url = {https://github.com/nschloe/pygalmesh},
-	author = {Schlömer, Nico},
+	author = {Schlömer, Nico, et al.},
 	year = {2022},
 	doi = {10.5281/zenodo.5564818},
 }
@@ -430,12 +430,21 @@ Section: New Results},
 	title = {{CalculiX}: {A} {Three}-{Dimensional} {Structural} {Finite} {Element} {Program}},
 	url = {http://www.dhondt.de/},
 	urldate = {2022-08-14},
-	author = {Dhont, Guido},
+	author = {Dhont, Guido, et al.},
 	year = {2022},
 }
 
+@book{dhondt2004finite,
+  title={The finite element method for three-dimensional thermomechanical applications},
+  author={Dhondt, Guido, et al.},
+  url = {http://www.dhondt.de/},
+  year={2004},
+  publisher={John Wiley \& Sons}
+}
+
 @misc{noauthor_paraview_2022,
-	title = {{ParaView}},
+	title = {{ParaView}, A Parallel Visualization Application. Kitware Inc.},
+	author = {A. Henderson, et al.},
 	url = {https://www.paraview.org/},
 	language = {en-US},
 	urldate = {2022-08-14},
@@ -459,17 +468,17 @@ Section: New Results},
 	month = nov,
 	year = {2014},
 	note = {Number: 6
-Publisher: International Union of Crystallography},
+	Publisher: International Union of Crystallography},
 	pages = {1224--1230},
 }
 
 @misc{noauthor_calculix_2022,
 	title = {{CalculiX} to {Paraview} converter},
 	copyright = {GPL-3.0},
+	author = {Ihor Mirzov, et al.},
 	url = {https://github.com/calculix/ccx2paraview},
 	abstract = {CalculiX to Paraview converter (frd to vtk/vtu). Makes possible to view and postprocess CalculiX analysis results in Paraview. Generates Mises and Principal components for stress and strain tensors.},
 	urldate = {2022-08-14},
-	publisher = {CALCULIX},
 	year = {2022},
 	note = {original-date: 2019-05-10T20:03:23Z},
 	keywords = {calculix, converter, frd, paraview, vtk, vtu},
@@ -547,7 +556,7 @@ Publisher: International Union of Crystallography},
 	url = {https://github.com/pmneila/PyMCubes},
 	abstract = {Marching cubes (and related tools) for Python},
 	urldate = {2022-08-19},
-	author = {Neila, P. M.},
+	author = {Neila, P. M., et al.},
 	year = {2022},
 	note = {original-date: 2013-11-10T07:20:37Z},
 }
@@ -567,6 +576,7 @@ Publisher: International Union of Crystallography},
 
 @misc{noauthor_calculix_2022-1,
 	title = {{CalculiX} examples},
+	author = {Martin Kraska, et al.},
 	url = {https://github.com/calculix/examples},
 	abstract = {Official and user examples for CCX and CGX. All in one place. Do not hesitate to add yours.},
 	urldate = {2022-08-19},
