+++ JOSS_2022/paper.bib
@@ -413,7 +413,7 @@ Section: New Results},
 @misc{schlomer_meshio_2022,
 	title = {meshio v5.0.0: {Tools} for mesh files},
 	url = {https://github.com/nschloe/meshio},
-	author = {Schlömer, Nico, et al.},
+	author = {Schlömer, Nico},
 	year = {2022},
 	doi = {10.5281/zenodo.1173115},
 }
@@ -421,7 +421,7 @@ Section: New Results},
 @misc{schlomer_pygalmesh_2022,
 	title = {pygalmesh: {Python} interface for {CGAL}'s meshing tools},
 	url = {https://github.com/nschloe/pygalmesh},
-	author = {Schlömer, Nico, et al.},
+	author = {Schlömer, Nico},
 	year = {2022},
 	doi = {10.5281/zenodo.5564818},
 }
@@ -430,13 +430,13 @@ Section: New Results},
 	title = {{CalculiX}: {A} {Three}-{Dimensional} {Structural} {Finite} {Element} {Program}},
 	url = {http://www.dhondt.de/},
 	urldate = {2022-08-14},
-	author = {Dhont, Guido, et al.},
+	author = {Dhont, Guido},
 	year = {2022},
 }
 
 @book{dhondt2004finite,
   title={The finite element method for three-dimensional thermomechanical applications},
-  author={Dhondt, Guido, et al.},
+  author={Dhondt, Guido},
   url = {http://www.dhondt.de/},
   year={2004},
   publisher={John Wiley \& Sons}
@@ -444,7 +444,7 @@ Section: New Results},
 
 @misc{noauthor_paraview_2022,
 	title = {{ParaView}, A Parallel Visualization Application. Kitware Inc.},
-	author = {A. Henderson, et al.},
+	author = {A. Henderson},
 	url = {https://www.paraview.org/},
 	language = {en-US},
 	urldate = {2022-08-14},
@@ -475,7 +475,7 @@ Section: New Results},
 @misc{noauthor_calculix_2022,
 	title = {{CalculiX} to {Paraview} converter},
 	copyright = {GPL-3.0},
-	author = {Ihor Mirzov, et al.},
+	author = {Ihor Mirzov},
 	url = {https://github.com/calculix/ccx2paraview},
 	abstract = {CalculiX to Paraview converter (frd to vtk/vtu). Makes possible to view and postprocess CalculiX analysis results in Paraview. Generates Mises and Principal components for stress and strain tensors.},
 	urldate = {2022-08-14},
@@ -556,7 +556,7 @@ Section: New Results},
 	url = {https://github.com/pmneila/PyMCubes},
 	abstract = {Marching cubes (and related tools) for Python},
 	urldate = {2022-08-19},
-	author = {Neila, P. M., et al.},
+	author = {Neila, P. M.},
 	year = {2022},
 	note = {original-date: 2013-11-10T07:20:37Z},
 }
@@ -576,7 +576,7 @@ Section: New Results},
 
 @misc{noauthor_calculix_2022-1,
 	title = {{CalculiX} examples},
-	author = {Martin Kraska, et al.},
+	author = {Kraska, Martin},
 	url = {https://github.com/calculix/examples},
 	abstract = {Official and user examples for CCX and CGX. All in one place. Do not hesitate to add yours.},
 	urldate = {2022-08-19},
