+++ JOSS_2022/paper.bib
@@ -646,6 +646,7 @@ Section: New Results},
 
 @misc{LHDL,
 	title = {The {Living} {Human} {Digital} {Library}},
+	shorttitle = {LHDL},
 	url = {http://projects.kmi.open.ac.uk/lhdl/},
 	urldate = {2023-02-20},
 }
