+++ JOSS_2022/paper.bib
@@ -645,7 +645,8 @@ Section: New Results},
 
 
 @misc{LHDL,
-	title = {The {Living} {Human} {Digital} {Library}},
+	title = {The {Living} {Human} {Digital} {Library} (LHDL)},
+	year = {2006},
 	shorttitle = {LHDL},
 	url = {http://projects.kmi.open.ac.uk/lhdl/},
 	urldate = {2023-02-20},
