### 2023-02-20 gianthk/ciclope#16 @gianthk
@danasolav please check the latest version of the paper vs your list of comments/issues above. if you agree with all our replies please close this issue ;)
