### 2023-02-20 #4952 @gianthk
@danasolav please see all our replies to your open [issue](https://github.com/gianthk/ciclope/issues/16#issuecomment-1437078127) and let us know if you have more corrections/suggestions
