+++ JOSS_2022/paper.bib
@@ -561,6 +561,14 @@ Section: New Results},
 	note = {original-date: 2013-11-10T07:20:37Z},
 }
 
+@misc{pymcubes,
+	author = {{Pablo M\'{a}rquez Neila}},
+	title = {{PyMCubes}: marching cubes for {P}ython},
+	url = {https://github.com/pmneila/PyMCubes},
+	note = {\texttt{https://github.com/pmneila/PyMCubes}, version 0.1.2},
+	year = {2023},
+}
+
 @misc{noauthor_itkwidgets_2022,
 	title = {itkwidgets},
 	copyright = {Apache-2.0},
