+++ JOSS_2022/paper.md
@@ -77,7 +77,7 @@ A pipeline for the generation and solution of a FE model derived from 3D microCT
 
 - Voxel and tetrahedra mesh exports performed with `meshio` [@schlomer_meshio_2022].
 - Tetrahedra meshes generated with the Python [`CGAL`](https://www.cgal.org/) frontend `pygalmesh` [@schlomer_pygalmesh_2022].
-- High-resolution surface meshes generated with `PyMCubes` [@neila_pymcubes_2022].
+- High-resolution surface meshes generated with `PyMCubes` [@pymcubes].
 - FE input files (`.INP`) generated by `ciclope` can be solved using the free software `CalculiX` [@dhondt2004finite] or [`Abaqus`](https://www.3ds.com/products-services/simulia/products/abaqus/).
 - 3D images and FE results can be visualized with `itkwidgets` [@noauthor_itkwidgets_2022], `ParaView` [@noauthor_paraview_2022], and `ccx2paraview` [@noauthor_calculix_2022] as illustrated in the [example Jupyter notebooks](#Examples).
 
