+++ JOSS_2022/paper.bib
@@ -562,8 +562,8 @@ Section: New Results},
 }
 
 @misc{pymcubes,
-	author = {{Pablo M\'{a}rquez Neila}},
-	title = {{PyMCubes}: marching cubes for {P}ython},
+	author = {M\'{a}rquez Neila, Pablo},
+	title = {{PyMCubes}: marching cubes for {P}ython, v0.1.2},
 	url = {https://github.com/pmneila/PyMCubes},
 	note = {\texttt{https://github.com/pmneila/PyMCubes}, version 0.1.2},
 	year = {2023},
