### 2023-03-08 #4952 @gianthk
@danasolav did you have the time to review our replies to your open [issue](https://github.com/gianthk/ciclope/issues/16#issuecomment-1437078127)?
