+++ JOSS_2022/paper.bib
@@ -377,6 +377,7 @@ Section: New Results},
 	volume = {9},
 	number = {1},
 	journal = {IEEE Transactions on systems, man, and cybernetics},
+	doi = {10.1109/tsmc.1979.4310076},
 	author = {Otsu, Nobuyuki},
 	year = {1979},
 	pages = {62--66},
@@ -387,6 +388,7 @@ Section: New Results},
 	volume = {8},
 	number = {8},
 	journal = {IEEE trans syst Man Cybern},
+	doi = {10.1109/tsmc.1978.4310039},
 	author = {Ridler, TW and Calvard, S and {others}},
 	year = {1978},
 	pages = {630--632},
@@ -640,17 +642,6 @@ Section: New Results},
 	pages = {e453},
 	file = {Full Text PDF:/home/gianthk/Zotero/storage/HHLK6ZJK/Walt et al. - 2014 - scikit-image image processing in Python.pdf:application/pdf;Snapshot:/home/gianthk/Zotero/storage/S2SUMRD2/453.html:text/html},
 }
-
-@article{otsu_threshold_1979,
-	title = {A threshold selection method from gray-level histograms},
-	volume = {9},
-	number = {1},
-	journal = {IEEE Transactions on systems, man, and cybernetics},
-	author = {Otsu, Nobuyuki},
-	year = {1979},
-	pages = {62--66},
-}
-
 
 @misc{LHDL,
 	title = {The {Living} {Human} {Digital} {Library} (LHDL)},
