+++ src/ciclope/__init__.py
@@ -1,5 +1,5 @@
 """Computed Tomography to Finite Elements"""
-__version__ = "1.3.3"
+__version__ = "1.3.4"
 
 from .core import voxelFE, tetraFE
 # from .utils import preprocess, postprocess
