### 2023-03-10 #4952 @gianthk
@danasolav the new release is [v1.3.4](https://github.com/gianthk/ciclope/tree/v1.3.4).
Zenodo DOI: https://doi.org/10.5281/zenodo.7716943
