### 2023-03-13 #4952 @arfon
@danasolav @gianthk – it looks like the paper is referencing a figure that doesn't exist which is causing the JATS validation to fail (https://github.com/openjournals/joss-reviews/issues/4952#issuecomment-1465576720). Can you please see if you can resolve this?
