### 2023-03-13 #4952 @gianthk
@arfon @danasolav the issue was not a missing figure, but a wrong internal ref to the Examples section. It should be fixed now, pointing to the readthedocs Examples page.
