### 2023-03-25 #4952 @gianthk
> On the paper:
* [X]  Please add the town/city for affiliation 1
* [X]  Please write United Kingdom, instead of UK for affiliation 3.
On the archive:
* [X]  Please manually edit the archive listed license to match your open source license. It currently says "Creative Commons Attribution 4.0 International", but it should refer to the MIT license you use.
* [ ]  You may also edit the archive title to have the same use of capitals as the paper title e.g. : "Ciclope: micro Computed Tomography to Finite Elements"
> I can't find the difference in the paper and archive title..

* [X]  The author `Giulia Fraterrigo` appears missing from the archived version
* [ ]  You may want to update the "Publish in" field, which says: `Journal of Open Source Software (JOSS):.` but should perhaps say `Journal of Open Source Software (JOSS)`. i.e. remove `:.`.
> `:.` is there because journal volume and issue fields are empty. Once published, and if available, I can update the fields on the archive
