### 2023-03-29 c5978ad
fixed bug causing rectangular base models to fail
