### 2023-03-29 d90b939
fixed bug causing rectangular base models to fail
